feat: Implementacao controle de estoque + Ajustes.

+++ src/main/java/com/fiap/pj/core/orcamento/usecase/command/OrcamentoItemPecaInsumoCommand.java
@@ -3,13 +3,12 @@ package com.fiap.pj.core.orcamento.usecase.command;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 public record OrcamentoItemPecaInsumoCommand(
         @NotNull(message = "Identificador da peça/insumo não pode ser nulo")
         UUID pecaInsumoId,
         @PositiveOrZero(message = "Quantidade não pode ser negativa")
-        BigDecimal quantidade
+        Integer quantidade
 ) {
 }
