+++ src/main/resources/db/migration/V1.06__create_table_orcamentos_pecas_insumos.sql
@@ -5,11 +5,7 @@ CREATE TABLE orcamentos_pecas_insumos
     orcamento_id     UUID ,
     descricao        VARCHAR(255)  NOT NULL,
     preco            NUMERIC(10,2) NOT NULL,
-    quantidade       NUMERIC(10,2) NOT NULL,
+    quantidade       INTEGER NOT NULL,
     CONSTRAINT orcamentos_pecas_insumos_pkey  PRIMARY KEY (id),
     CONSTRAINT fk_orcamentos_pecas_insumos FOREIGN KEY (pecas_insumos_id) REFERENCES pecas_insumos(id)
 );
-
-
-
-
